restructuring files preparing for drills and removinglint-staged since it is redundant
