+++ sudokuru/app/Components/SudokuBoard/Core/Components/Cell.tsx
@@ -1,7 +1,7 @@
 import { getCellSize } from "../Functions/BoardFunctions";
 import React from "react";
 import { Pressable, StyleProp, Text, TextStyle, View } from "react-native";
-import { CellType } from "../../../Functions/LocalDatabase";
+import { CellType } from "../../../../Functions/LocalDatabase";
 
 let fallbackHeight = 30;
